requirements.txt
numpy==1.11.1
Django==1.10.1
Pillow==3.3.0
h5py==2.6.0
nltk==3.2.1
scipy==0.17.1
channels==0.17.2
asgi-redis==0.14.1
pika==0.10.0
pyyaml==3.12
uwsgi==2.0.13.1
websocket-client==0.37.0
daphne==0.15.0
